requirements.txt
maxcdn~=0.0
cryptography~=1.8
PyYAML~=3.12
